pyproject.toml
[project]
name = "imedia"
version = "0.1.0"
description = "Reddit → Medallion (raw/bronze/silver/gold) con Polars + SQLite"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "praw>=7.8.0",
  "polars>=1.5.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = ["ruff>=0.5.0","pytest>=8.0.0"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
